refactor: Create new views folder for view components
