+++ packages/frontend/src/hooks/useWorkouts.ts
@@ -2,9 +2,9 @@ import { useQuery, UseQueryResult } from "react-query";
 import { useRecoilValue } from "recoil";
 
 import { getWorkouts } from "../apis/workoutsAPI";
-import { dateState } from "../Daily/states";
 import { IWorkout } from "../types";
 import { getDateString } from '../utils/getDateString';
+import { dateState } from "../views/Daily/states";
 
 export function useWorkouts(): UseQueryResult<IWorkout[], Error> {
   const date = useRecoilValue(dateState);
