+++ packages/frontend/src/views/Daily/SelectRoutine/RoutineItem.tsx
@@ -3,9 +3,9 @@ import React from 'react';
 import { useQueryClient } from 'react-query';
 import { useRecoilValue } from 'recoil';
 
-import { createWorkoutsByRoutine } from '../../apis/workoutsAPI';
-import { IRoutine } from '../../types';
-import { getDateString } from '../../utils/getDateString';
+import { createWorkoutsByRoutine } from '../../../apis/workoutsAPI';
+import { IRoutine } from '../../../types';
+import { getDateString } from '../../../utils/getDateString';
 import { dateState } from '../states';
 import ExerciseItem from './ExerciseItem';
