+++ packages/frontend/src/views/Daily/components/Workout.tsx
@@ -9,8 +9,8 @@ import {
 } from '@material-ui/icons';
 import React, { useState } from 'react';
 
-import { updateDoneCount } from '../../apis/workoutsAPI';
-import { IWorkout } from '../../types';
+import { updateDoneCount } from '../../../apis/workoutsAPI';
+import { IWorkout } from '../../../types';
 
 const useStyles = makeStyles({
   fadeInOutContainer: {
