+++ packages/frontend/src/views/Daily/index.tsx
