+++ packages/frontend/src/views/Daily/states.ts
